Added Poloniex wallet out of js.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/lib/ui.h
@@ -11,8 +11,8 @@ namespace K {
   uWS::Group<uWS::SERVER> *uiGroup = hub.createGroup<uWS::SERVER>(uWS::PERMESSAGE_DEFLATE);
   int iOSR60 = 0;
   bool uiOPT = true;
-  double uiMDT = 0;
-  double uiDDT = 0;
+  unsigned long uiMDT = 0;
+  unsigned long uiDDT = 0;
   string uiNOTE = "";
   string uiNK64 = "";
   Persistent<Function> socket_;
